feat: code optimisation, bug fixes, introduction of new features
Rewrote the game functions, now it is one function with innovations. Now you need to pass 3 values into the function: max_number, max_tries, max_time; Now every time a message with the first hint is displayed, no additional maths is done; If max_time and/or max_tries is equal to 0, the program will set ‘infinite’ values; renamed the name of the function with time countdown to a more understandable one; added that if time or attempts are infinite, their number will not be displayed to the user; added error handling about wrong data type in user_guess; added program termination if the game ended with a loss due to exhaustion of attempts.

Updated logic for custom_level function. Now there is an error handler there because of entered incorrect data types.

Bug fixes: fixed a bug where the programme would close if the user entered a random set of letters/numbers when selecting a difficulty; fixed a bug where when the game ended the programme did not terminate; fixed a bug where the end of time did not mean the end of the programme; fixed a bug where the message with the exhausted time was shown immediately after the colon when the user entered a number.